requirements.txt
flax>=0.7.2
ray[tune]
jax>=0.4.14
jraph>=0.0.6.dev0
jaxtyping>=0.2.21
numpy>=1.25.2,<2
einops>=0.6.1
matplotlib>=3.7.2
opencv-python>=4.8.0.76
tqdm>=4.66.1
tensorflow-probability>=0.21.0
optax>=0.1.7
scipy>=1.11.2
wandb>=0.15.8
pyyaml>=6.0.1
orbax_checkpoint>=0.3.5
seaborn>=0.12.2
equinox>=0.11.0
loguru>=0.7.2
attrs>=23.1.0
rich>=13.5.3
ipdb>=0.13.13
colour>=0.1.5
control>=0.9.4
h5py >= 3.11.0
distrax >= 0.1.3
cython<3
d4rl @ git+https://github.com/Farama-Foundation/d4rl@master#egg=d4rl
diff-spec @ git+https://github.com/jeappen/diff-spec.git@feature/jax
jaxproxqp @ git+https://github.com/oswinso/jaxproxqp.git